Split animations.rs into smaller files
